requirementsV2.txt
absl-py==2.1.0
aiofiles==23.2.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.11
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.7.0
anyio==4.4.0
asttokens==2.4.1
astunparse==1.6.3
async-timeout==5.0.1
attrs==23.2.0
backcall==0.2.0
blinker==1.8.2
cachetools==5.5.0
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
ConfigArgParse==1.7
contourpy==1.1.1
cycler==0.12.1
Cython==3.0.10
dash==2.18.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
decorator==5.1.1
dnspython==2.6.1
email_validator==2.1.1
exceptiongroup==1.2.1
executing==2.1.0
fastapi==0.111.0
fastapi-cli==0.0.4
fastjsonschema==2.20.0
ffmpy==0.3.2
filelock==3.14.0
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.53.0
freetype-py==2.5.1
frozenlist==1.5.0
fsspec==2024.6.0
gast==0.4.0
glfw==2.8.0
google-auth==2.37.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
gradio==4.32.2
gradio_client==0.17.0
grpcio==1.69.0
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.2
humanfriendly==10.0
idna==3.7
imageio==2.35.1
imageio-ffmpeg==0.5.1
importlib_metadata==8.1.0
importlib_resources==6.4.0
intel-openmp==2021.4.0
ipython==8.12.3
ipywidgets==8.1.5
itsdangerous==2.2.0
jax==0.4.13
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
keras==2.13.1
kiwisolver==1.4.5
lazy_loader==0.4
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.7
mdurl==0.1.2
mediapipe==0.10.11
mkl==2021.4.0
ml-dtypes==0.2.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.1
numpy==1.24.0
oauthlib==3.2.2
OneEuroFilter==0.2.1
onnx==1.16.1
onnxruntime-gpu==1.19.2
open3d==0.19.0
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.82
opt-einsum==3.3.0
orjson==3.10.3
packaging==24.0
pandas==2.0.3
parso==0.8.4
pickleshare==0.7.5
pillow==10.3.0
pkgutil_resolve_name==1.3.10
platformdirs==4.3.6
plotly==5.24.1
pooch==1.8.2
prompt_toolkit==3.0.48
propcache==0.2.0
protobuf==3.20.3
psutil==6.0.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pybind11==2.13.6
pycparser==2.22
pydantic==2.7.3
pydantic_core==2.18.4
pydub==0.25.1
pyglet==2.1.6
Pygments==2.18.0
pymeshfix==0.16.3
PyOpenGL==3.1.0
pyparsing==3.1.2
pyreadline3==3.4.1
pyrealsense2==2.55.1.6486
pyrender==0.1.45
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
pyvista==0.44.2
PyWavelets==1.4.1
pywin32==308
PyYAML==6.0.1
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
retrying==1.3.4
rich==13.7.1
rpds-py==0.18.1
rsa==4.9
ruff==0.4.7
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
scooby==0.10.0
semantic-version==2.10.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sounddevice==0.4.7
stack-data==0.6.3
starlette==0.37.2
sympy==1.12.1
tbb==2021.12.0
tenacity==9.0.0
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-intel==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
threadpoolctl==3.5.0
tifffile==2023.7.10
tomlkit==0.12.0
toolz==0.12.1
torch==2.4.1+cu124
torchaudio==2.4.1+cu124
torchvision==0.19.1+cu124
tqdm==4.66.4
traitlets==5.14.3
transforms3d==0.4.2
trimesh==4.5.1
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
urllib3==2.2.1
uvicorn==0.30.1
vtk==9.3.1
watchfiles==0.22.0
wcwidth==0.2.13
websockets==11.0.3
Werkzeug==3.0.6
widgetsnbextension==4.0.13
wrapt==1.17.2
wslink==2.2.1
yarl==1.15.2
zipp==3.19.1
